+++ src/analysis/analysisVisitors/eventNameAnalysis/checkEventNamesAnalysis.ts
@@ -44,13 +44,9 @@ class CheckEventNamesAnalysis extends AnalysisVisitorABC {
 
   get knownEventNames(): string[] {
     return [
-      ...this.collectedEvents.knownNames,
-      ...this.collectedEvents.triggerNames,
+      ...this.collectedEvents.knownEmittedNames,
+      ...this.collectedEvents.knownTriggerNames,
     ];
-  }
-
-  get knownTriggerEventNames(): string[] {
-    return this.collectedEvents.triggerNames;
   }
 
   override visitExtensionPoint(
@@ -73,7 +69,7 @@ class CheckEventNamesAnalysis extends AnalysisVisitorABC {
         });
       } else if (
         !DOM_EVENTS.includes(eventName) &&
-        !this.collectedEvents.knownNames.includes(
+        !this.collectedEvents.knownEmittedNames.includes(
           extensionPoint.definition.customEvent.eventName
         )
       ) {
@@ -108,8 +104,12 @@ class CheckEventNamesAnalysis extends AnalysisVisitorABC {
     );
 
     this.collectedEvents = {
-      knownNames: uniq(flatten(results.map((result) => result.knownNames))),
-      triggerNames: uniq(flatten(results.map((result) => result.triggerNames))),
+      knownEmittedNames: uniq(
+        flatten(results.map((result) => result.knownEmittedNames))
+      ),
+      knownTriggerNames: uniq(
+        flatten(results.map((result) => result.knownTriggerNames))
+      ),
       hasDynamicEventName: results.some((result) => result.hasDynamicEventName),
     };
