+++ src/bricks/PipelineVisitor.ts
@@ -25,7 +25,7 @@ import {
   getRootPipelineFlavor,
   getSubPipelineFlavor,
 } from "@/bricks/blockFilterHelpers";
-import { type StarterBrickType } from "@/starterBricks/types";
+import { type StarterBrickType } from "@/starterBricks/StarterBrickType";
 import { PipelineFlavor } from "@/pageEditor/pageEditorTypes";
 import { PIPELINE_BLOCKS_FIELD_NAME } from "@/pageEditor/consts";
 import { isPipelineExpression } from "@/utils/expressionUtils";
