+++ src/starterBricks/starterBrickModUtils.ts
@@ -19,10 +19,8 @@ import {
   type ModComponentDefinition,
   type ModDefinition,
 } from "@/types/modDefinitionTypes";
-import {
-  type StarterBrickDefinition,
-  type StarterBrickType,
-} from "@/starterBricks/types";
+import { type StarterBrickDefinition } from "@/starterBricks/types";
+import { type StarterBrickType } from "./StarterBrickType";
 import starterBrickRegistry from "@/starterBricks/registry";
 import { type RegistryId } from "@/types/registryTypes";
 import { compact, uniq } from "lodash";
