+++ src/testUtils/factories/modDefinitionFactories.ts
@@ -38,9 +38,9 @@ import {
 } from "@/types/helpers";
 import {
   type StarterBrickConfig,
-  type StarterBrickType,
   type StarterBrickDefinition,
 } from "@/starterBricks/types";
+import { type StarterBrickType } from "@/starterBricks/StarterBrickType";
 import { DEFAULT_EXTENSION_POINT_VAR } from "@/pageEditor/starterBricks/base";
 import { type SafeString } from "@/types/stringTypes";
