+++ src/testUtils/factories/pageEditorFactories.ts
@@ -31,10 +31,8 @@ import {
 } from "@/types/runtimeTypes";
 import { uuidSequence } from "@/testUtils/factories/stringFactories";
 import { type IntegrationDependency } from "@/types/integrationTypes";
-import {
-  type StarterBrickConfig,
-  type StarterBrickType,
-} from "@/starterBricks/types";
+import { type StarterBrickConfig } from "@/starterBricks/types";
+import { type StarterBrickType } from "@/starterBricks/StarterBrickType";
 import {
   starterBrickConfigFactory,
   metadataFactory,
